Refactor BootstrapConfig (#698)
* wip

* refactor bootstrap config

* fix broken tests:

* fix failing cli build

* improve readability of bootstrap config changes

* add bootstrap config back into test helper so tests pass

---------
