+++ crates/cli/tests/whitelist_data/whitelist.json
@@ -1,1 +1,1 @@
-{"genesis-farmers":[{"kademlia_liveness_address":"127.0.0.1:13004","kademlia_peer_id":[142,7,26,22,211,7,23,130,240,218,96,15,104,205,68,72,243,204,62,33,184,44,222,254,190,50,110,217,164,198,128,168],"node_id":"node-4","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12004","udp_gossip_address":"127.0.0.1:11004","validator_public_key":"0275a4c05daee734fe2525849e235a00cf982db298a8c2d75983e31d9f2f11792f"},{"kademlia_liveness_address":"127.0.0.1:13005","kademlia_peer_id":[25,154,72,81,179,51,139,156,154,170,94,174,86,42,198,17,173,151,34,156,80,80,107,138,167,64,226,222,132,67,24,137],"node_id":"node-5","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12005","udp_gossip_address":"127.0.0.1:11005","validator_public_key":"0345ac2c7c5b5273b243ed2d8377d3e7e5cfd07c30523d290f32fcf25874e6208c"}],"genesis-harvesters":[{"kademlia_liveness_address":"127.0.0.1:13002","kademlia_peer_id":[190,137,211,36,155,123,219,247,149,93,186,222,179,179,74,153,210,137,136,19,200,176,208,164,255,45,128,19,45,161,60,17],"node_id":"node-2","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12002","udp_gossip_address":"127.0.0.1:11002","validator_public_key":"024e8b3c3721c25328b86575e7240e8fdf2110614a0a6b3e6d93e7bdd5278c990a"},{"kademlia_liveness_address":"127.0.0.1:13003","kademlia_peer_id":[56,165,111,28,33,248,122,201,94,247,159,19,222,198,232,75,222,11,220,75,250,118,44,231,147,225,117,55,241,144,154,60],"node_id":"node-3","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12003","udp_gossip_address":"127.0.0.1:11003","validator_public_key":"03daeca234a50561fa415c9f79f1a15095702f9a085ae4915282ed775971163f75"}],"genesis-miner":{"kademlia_liveness_address":"127.0.0.1:13001","kademlia_peer_id":[190,187,253,208,220,245,39,153,57,89,198,92,151,36,251,120,95,206,239,235,246,178,132,162,45,89,217,143,126,241,80,254],"node_id":"node-1","node_type":"bootstrap","raptorq_gossip_address":"127.0.0.1:12001","udp_gossip_address":"127.0.0.1:11001","validator_public_key":"02cefe78bf8b776d890e05219957be64d8b5de318f0d152d9b34105890ed3aa712"}}
+{"genesis-farmers":[{"kademlia_liveness_address":"127.0.0.1:13004","kademlia_peer_id":[117,126,103,73,226,157,146,198,67,97,109,239,8,251,100,123,129,0,64,206,244,40,217,194,252,86,85,147,173,219,216,135],"node_id":"node-4","node_type":"validator","quorum_kind":"Harvester","raptorq_gossip_address":"127.0.0.1:12004","udp_gossip_address":"127.0.0.1:11004","validator_public_key":"03325bccc87803b9082e5d8c5f050361cb0b703a7dfc960a2a7a5abd8911c62cda"},{"kademlia_liveness_address":"127.0.0.1:13005","kademlia_peer_id":[121,147,191,79,218,70,196,205,13,5,240,222,42,221,189,9,173,238,110,50,34,36,112,64,153,66,101,103,110,62,219,151],"node_id":"node-5","node_type":"validator","quorum_kind":"Harvester","raptorq_gossip_address":"127.0.0.1:12005","udp_gossip_address":"127.0.0.1:11005","validator_public_key":"02e11536b6f748adee8023fc9c2ec13c20958ae761600f10ea3e06ecac29479534"}],"genesis-harvesters":[{"kademlia_liveness_address":"127.0.0.1:13002","kademlia_peer_id":[151,42,206,118,10,71,65,223,177,109,187,185,219,53,130,34,207,167,93,169,198,54,68,181,165,182,175,53,158,85,181,66],"node_id":"node-2","node_type":"validator","quorum_kind":"Harvester","raptorq_gossip_address":"127.0.0.1:12002","udp_gossip_address":"127.0.0.1:11002","validator_public_key":"0273cea7d96613fed60c054f9f5bfdbb02e8c947c9d38b115d29c8b27efd2eb747"},{"kademlia_liveness_address":"127.0.0.1:13003","kademlia_peer_id":[11,9,106,144,133,33,14,36,202,30,90,255,121,33,119,65,34,68,113,25,142,119,227,67,70,89,163,157,153,22,104,221],"node_id":"node-3","node_type":"validator","quorum_kind":"Harvester","raptorq_gossip_address":"127.0.0.1:12003","udp_gossip_address":"127.0.0.1:11003","validator_public_key":"0241ec799d3f0638efd9152727e81cf1d2f3d2397ab746761acd08d99545088bf0"}],"genesis-miner":{"kademlia_liveness_address":"127.0.0.1:13001","kademlia_peer_id":[104,161,112,111,104,21,109,86,213,56,226,32,100,11,125,43,63,245,35,98,189,137,11,63,34,61,175,67,55,29,17,198],"node_id":"node-1","node_type":"bootstrap","quorum_kind":"Harvester","raptorq_gossip_address":"127.0.0.1:12001","udp_gossip_address":"127.0.0.1:11001","validator_public_key":"02f4efceff974b37b0206c744f315dd8252137a3f57af4295dcd4c5c538a231e49"}}
