+++ crates/node/src/runtime/node_runtime_handler.rs
@@ -75,7 +75,7 @@ impl Handler<EventMessage> for NodeRuntime {
                             })
                             .collect();
 
-                        //add the addresses from config.bootstrap_config.additional_genesis_receivers to the genesis_receivers list
+                        // add the addresses from config.bootstrap_config.additional_genesis_receivers to the genesis_receivers list
                         if let Some(bootstrap_config) = &self.config.bootstrap_config {
                             if let Some(additional_genesis_receivers) =
                                 &bootstrap_config.additional_genesis_receivers
