+++ crates/vrrb_rpc/src/rpc/api.rs
@@ -7,7 +7,7 @@ use primitives::{Address, NodeType, Round};
 use secp256k1::PublicKey;
 use serde::{Deserialize, Serialize};
 use storage::vrrbdb::Claims;
-use vrrb_config::bootstrap_quorum::QuorumMembershipConfig;
+use vrrb_config::QuorumMembershipConfig;
 use vrrb_core::account::Account;
 use vrrb_core::node_health_report::NodeHealthReport;
 use vrrb_core::transactions::{
