+++ examples/multichain.rs
@@ -1,7 +1,7 @@
-use std::str::FromStr;
 use node::test_utils::create_test_network;
-use telemetry::custom_subscriber::TelemetrySubscriber;
 use primitives::{Address, PublicKey};
+use std::str::FromStr;
+use telemetry::custom_subscriber::TelemetrySubscriber;
 
 #[tokio::main]
 async fn main() {
